+++ lib/shale/mapping/dict.rb
@@ -5,22 +5,32 @@ require_relative 'dict_group'
 
 module Shale
   module Mapping
-    # Mapping for dictionary serialization formats (Hash/JSON/YAML)
+    # Mapping for dictionary serialization formats (Hash/JSON/YAML/TOML/CSV)
     #
     # @api private
     class Dict < DictBase
       # Map key to attribute
       #
       # @param [String] key Document's key
-      # @param [Symbol, nil] to Object's attribute
+      # @param [Symbol, nil] to
+      # @param [Symbol, nil] receiver
       # @param [Hash, nil] using
-      # @param [true, false] render_nil
+      # @param [true, false, nil] render_nil
       #
       # @raise [IncorrectMappingArgumentsError] when arguments are incorrect
       #
       # @api private
-      def map(key, to: nil, using: nil, render_nil: false)
-        super(key, to: to, using: using, render_nil: render_nil)
+      def map(key, to: nil, receiver: nil, using: nil, render_nil: nil)
+        super(key, to: to, receiver: receiver, using: using, render_nil: render_nil)
+      end
+
+      # Set render_nil default
+      #
+      # @param [true, false] val
+      #
+      # @api private
+      def render_nil(val)
+        @render_nil_default = val
       end
 
       # Map group of keys to mapping methods