+++ spec/shale/type/value_spec.rb
@@ -59,9 +59,21 @@ RSpec.describe Shale::Type::Value do
     end
   end
 
+  describe '.of_csv' do
+    it 'returns value' do
+      expect(described_class.of_csv(123)).to eq(123)
+    end
+  end
+
+  describe '.as_csv' do
+    it 'returns value' do
+      expect(described_class.as_csv(123)).to eq(123)
+    end
+  end
+
   describe '.of_xml' do
     it 'extracts text from XML node' do
-      element = ::REXML::Element.new('name')
+      element = REXML::Element.new('name')
       element.add_text('foobar')
 
       node = Shale::Adapter::REXML::Node.new(element)